tessera/requirements.txt
# Core dependencies
pyjwt==2.8.0
cryptography==42.0.5
python-dotenv==1.0.0
streamlit==1.32.0
plotly==5.19.0
pandas==2.2.0
requests==2.31.0
prometheus-client==0.20.0
ldap3==2.9.1
python3-saml==1.16.0
python-multipart==0.0.9

# Optional dependencies (commented out for now)
# hvac==2.1.0  # HashiCorp Vault
# boto3==1.34.0  # AWS KMS
# redis==5.0.1  # Redis client
# psycopg2-binary==2.9.9  # PostgreSQL

# Development
pytest==8.0.0
black==24.1.1
# Production dependencies
redis==5.0.1
streamlit-autorefresh==1.0.1
